// guig/shared/types.ts
export type FileState =
  | "modified"
  | "added"
  | "deleted"
  | "renamed"
  | "copied"
  | "untracked"
  | "conflicted";

export interface ChangedFile {
  path: string;
  originalPath?: string;
  state: FileState;
  staged: boolean;
  unstaged: boolean;
  submodule?: string;
}